// frontend/src/app/api-services/api/model/group-response.ts
/**
 * StudyConnect Backend API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface GroupResponse {
    id: number;
    name: string;
    description?: string;
    is_public: boolean;
    created_by_id: number;
    created_at: string;
    last_updated_at: string;
    member_count: number;
    max_members: number;
    invite_code: string;
}